requirements.txt
aicspylibczi==3.0.4
Bio==1.5.9
bioformats==0.1.15
biopython==1.79
czifile==2019.7.2
h5py==3.6.0
javabridge==1.0.19
keras==2.8.0
matplotlib==3.5.1
numpy==1.20.3
opencv_python==4.5.4.60
pandas==1.3.5
Pillow==9.5.0
python_bioformats==4.0.5
python_javabridge==4.0.3
scikit_image==0.19.1
scikit_learn==1.0.1
scipy==1.7.3
tensorflow_gpu==2.8.0
tifffile==2021.11.2
tqdm==4.65.0
